requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
boto3==1.34.19
botocore==1.34.19
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
dataclasses-json==0.5.14
dill==0.3.7
distro==1.9.0
Django==4.2.9
django-storages==1.14.2
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
exceptiongroup==1.2.0
faiss-cpu==1.7.4
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.189
langchain-community==0.0.12
langchain-core==0.1.10
langsmith==0.0.80
marshmallow==3.20.2
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.6
numpy==1.24.4
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.2
pydantic==1.10.13
pydantic-core==2.14.6
pymongo==4.6.1
pypdf2==3.0.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.66.1
typing-extensions==4.9.0
typing-inspect==0.9.0
tzdata==2023.4
urllib3==1.26.18
yarl==1.9.4
